requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
boto3>=1.20.0
pyyaml>=6.0
plotly>=5.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
requests>=2.26.0